refactor: move translateTabStatus function to the utils and use it inside BookingModal
